neutrinoFilterGoldenSample.cxx no longer crashes when MCTracks->GetEntries() == 0
MCEventBuilder is causing some MC events to not have MCTracks. This crashed
the afforementioned selection file when filling histograms (checks for the
energy of the MCTrack in case we are looking at MC)
Added a clause to skip that step when MCTracks->GetEntries() == 0 and
a new histogram "No_MCTracks" that keeps track of how many events in
the MC file met this criteria.